NTS-9167: Use buffer to hold chunk data to attempt retries
Some of the content types are not seekable (eg response) so if we attempt to
retry then there won't be any data available to send. This patch introduces a
temporary buffer that holds the chunk data which can be read back during
retries.

There is a performance penalty in case of large number of chunks or large chunk
size hence this is only attempted if user requests for Chunk retries otherwise
existing behaviour is retained.